Add `install` task.
Task `install` installs the module into a temporary destination folder.

Added `FSDriver`, which acts as an interface to Ruby's filesystem API.

Added `Modulefile` class, which acts as an interface to library's main input
(that is, the module's Modulefile), parsing it into an useful data structure.

Added helper functions to improve tests readability.