Add `deb` task.
Task `deb` creates a Debian package out of the module source code; for the
moment it reads just few attributes from the `Modulefile`.

Renamed `FSDriver` to `System`.
Extracted parameter objects for `Install` and `Clean` classes so to match
interface of `Deb` class (not merged into a superclass yet, but sure if it's
worth something).