Add support for dependency parsing.
The library can now read all the dependencies from a Modulefile, and turn them
into proper FPM invocation options. A word worth saying: in the Modulefile,
dependencies are specified as `mod_name constraint_1 constraint_2 ...`, but
using the same format for the `-d` option doesn't work at least for RPMs. We
need instead to split every version constraint (like `<= 3.0.0`) and put it
into a separate `-d` option. This is why the code is a little more convoluted
than I'd have liked it to be.